Second Submit

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task5;
 public class NumberDivider {
     void divide(int toBeDivided) {
     
-	int result = toBeDivided / 5;
+	float result = toBeDivided / 5;
 	System.out.println(result);
